requirements.txt
Django>=3.2.3,<4.1
psycopg2>=2.8.6,<2.9
djangorestframework==3.12.2
django-cors-headers>=3.6.0,<3.8
Pillow
boto3==1.18.33
django-storages==1.11.1
uWSGI>=2.0.19.1,<2.1
pytz